added docs and migrate to TypeScript
